Create contact_types package
This commit creates a new package, contact_types. It contains:
* item.dart
    * MultaccItem is the base multacc item class
    * MultaccItemType is an enum of item types
* twitter.dart
    * TwitterItem is a class extending MultaccItem and demonstrating
      what the implementations of MultaccItem should look like.

In the future there should be one class for each item type.

The indices from MultaccItemType will be used to identify the item types
(Snapchat, Instagram, etc.) in the database.

This package should contain classes that extend MultaccItem for each
item type. A MultaccContact will then contain a List of instances of
these classes, which will also be persisted in databases as necessary.

MultaccItem is an abstract class; all items should be instances of
classes such as TwitterItem that extend MultaccItem.