Update MultaccItem to use json serialization
Previously, the only data stored besides the database key in a
MultaccItem was a `value` string. This commit modifies the MultaccItem
base class to replace this string with a `fromJson` constructor and
`toJson` method.

The TwitterItem class provides an example implementation. The user ID
and display name are stored in two instance variables in the TwitterItem
class. The JSON stored in the database will look like:
    {'at':'fckeveryword','id':'944864788336824321'}

This way, implementations of MultaccItem can store arbitrary data as
necessary for the given platform.