Store additional MultaccContact fields in database
Closes #70.

In this commit, only display name, avatar, and birthday are stored
alongside MultaccItems. Additional fields should be added as they
become necessary.

The following changes were made to facilitate this:
* Add Hive adapters for MultaccContact and MultaccItem
* Modify MultaccItem.fromDB to take only the JSON value as a parameter
    * Modify dummy contacts json to include key
* Create an empty default constructor for MultaccContact
* Name the existing MultaccContact(Contact) constructor as fromContact
* Remove the existing addContact method from the database interface and
  rename addMultaccContact to addContact
* Add names to the dummy contacts
* DatabaseInterface.getContact now returns null if the contact does
  not exist (previously it returned an empty list of items)

Other changes while I'm here:
* Add constants in item.dart for '_t' and '_id' magic keys
* Add toString to MultaccItem, which calls jsonEncode on it
* Add DatabaseInterface.getAllContacts
* Rename hive box from 'box' to 'contacts'
* Contacts box can now only contain contacts
* Move Hive box info from main.dart into DatabaseInterface
    * Makes adding additional boxes easier
    * Closes #67
* Rename DatabaseInterface instance in main from dbi to db

This uses a generated Hive adapter for MultaccContact. In order to
generate the adapter, we have to apply annotations to the fields we want
to store, which  means MultaccContact has to override the fields from
Contact we want to store. Overriding fields is generally frowned upon,
so if this is a concern the adapter can be replaced with a manually
written one in the future.

We cannot generate a Hive adapter for MultaccItem because it has several
implementations with their own fields that have to be stored. We *could*
generate adapters for each MultaccItem implementation, but then we would
need code when storing the items to detect their type and code when
reading them from the database to handle each type differently. For the
same reason, the adapter I have written continues to use our existing
JSON serialization for MultaccItem.

I created a file, type_ids.dart, which contains constants for the type
IDs for Hive adapters. This is because these must be globally unique,
and keeping track of them in a single file will be easier than having
them each in the file with the adapter.

As MultaccItem.fromDB no longer takes parameters other than a single
json-formatted string, it is conceivable to make it the fromJson method.
This would necessitate renaming the existing fromJson method to fromMap,
paralleling toMap; this is the method that would be separately implemen-
ted for each item type.