Generate display name for contacts without one
Closes #135.

This adds a name getter on MultaccContact that should be read instead of
displayName. If displayName is set, its value is returned. Otherwise, it
attempts to use the other name fields. If none of these contain a name,
it returns the human-readable value of the first item with one. If no
items have a human-readable value, it falls back to the text 'Contact'.